LPS-141119 Interface is not needed
