+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/init.jsp
@@ -46,7 +46,6 @@ page import="com.liferay.search.experiences.web.internal.blueprint.admin.constan
 page import="com.liferay.search.experiences.web.internal.blueprint.admin.display.context.ViewSXPBlueprintsDisplayContext" %><%@
 page import="com.liferay.search.experiences.web.internal.blueprint.admin.display.context.ViewSXPElementsDisplayContext" %><%@
 page import="com.liferay.search.experiences.web.internal.blueprint.options.portlet.preferences.SXPBlueprintOptionsPortletPreferences" %><%@
-page import="com.liferay.search.experiences.web.internal.blueprint.options.portlet.preferences.SXPBlueprintOptionsPortletPreferencesImpl" %><%@
 page import="com.liferay.search.experiences.web.internal.constants.SXPWebKeys" %>
 
 <liferay-frontend:defineObjects />
